+++ includes/rts/Flags.h
@@ -115,7 +115,7 @@ typedef struct _COST_CENTRE_FLAGS {
 # define COST_CENTRES_SUMMARY	1
 # define COST_CENTRES_VERBOSE	2 /* incl. serial time profile */
 # define COST_CENTRES_ALL	3
-# define COST_CENTRES_XML       4
+# define COST_CENTRES_JSON      4
 
     int	    profilerTicks;   /* derived */
     int	    msecsPerTick;    /* derived */
@@ -136,7 +136,7 @@ typedef struct _PROFILING_FLAGS {
 
     Time        heapProfileInterval; /* time between samples */
     uint32_t    heapProfileIntervalTicks; /* ticks between samples (derived) */
-    bool     includeTSOs;
+    bool        includeTSOs;
 
 
     bool		showCCSOnException;